coop_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
python-decouple>=3.8
requests>=2.31
python-docx>=1.1
Pillow>=10.0
